✨ Feat:#126-성능 개선
useClientSyncStore를 hook으로 빼서 관리
