+++ src/app/layout.tsx
@@ -1,5 +1,6 @@
 import type { Metadata } from 'next';
 import localFont from 'next/font/local';
+import ClientSyncWrapper from '@/components/ClientWrapper/ClientSync';
 import Nav from '@/components/Nav/Nav';
 import '../styles/globals.css';
 import ReactQueryProviders from '../../hooks/useReactQuery';
@@ -27,6 +28,7 @@ export default function RootLayout({
       </head>
       <body className={pretendard.className}>
         <ReactQueryProviders>
+          <ClientSyncWrapper />
           <Nav />
           <div className="min-h-screen bg-gray-50">{children}</div>
         </ReactQueryProviders>
