+++ src/components/ClientWrapper/ClientSync.tsx
@@ -1,0 +1,11 @@
+'use client';
+
+import { useClientSyncStore } from '../../../hooks/useClientSyncStore';
+
+const ClientSyncWrapper = () => {
+  useClientSyncStore();
+
+  return null;
+};
+
+export default ClientSyncWrapper;
